+++ airo-camera-toolkit/airo_camera_toolkit/cameras/realsense/realsense.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import time
 from typing import Any, Optional
 
+import cv2
 import numpy as np
 import pyrealsense2 as rs  # type: ignore
 from airo_camera_toolkit.interfaces import RGBDCamera
@@ -10,6 +11,7 @@ from airo_camera_toolkit.utils.image_converter import ImageConverter
 from airo_typing import (
     CameraIntrinsicsMatrixType,
     CameraResolutionType,
+    NumpyConfidenceMapType,
     NumpyDepthMapType,
     NumpyFloatImageType,
     NumpyIntImageType,
@@ -25,6 +27,7 @@ class Realsense(RGBDCamera):
     * Depth resolution is automatically set
     * Depth frames are always aligned to color frames
     * Hole filling is enabled by default
+    * Getting the confidence map is optional (disabled by default). The L515 supports a built-in confidence map, but the D415 and D435 do not. We do not support the L515 at this time.
     """
 
     # Built-in resolutions (16:9 aspect ratio) for convenience
@@ -39,12 +42,16 @@ class Realsense(RGBDCamera):
         resolution: CameraResolutionType = RESOLUTION_1080,
         fps: int = 30,
         enable_depth: bool = True,
+        enable_confidence_map: bool = False,
         enable_hole_filling: bool = True,
         serial_number: Optional[str] = None,
     ) -> None:
         self._resolution = resolution
         self._fps = fps
         self._depth_enabled = enable_depth
+        self._confidence_enabled = enable_confidence_map
+        if self._confidence_enabled and not self._depth_enabled:
+            raise ValueError("enable_confidence_map can only be True if enable_depth is also True")
         self.hole_filling_enabled = enable_hole_filling
         self.serial_number = serial_number
 
@@ -56,10 +63,14 @@ class Realsense(RGBDCamera):
 
         config.enable_stream(rs.stream.color, resolution[0], resolution[1], rs.format.rgb8, fps)
 
+        # Use max resolution that can handle the fps for depth (will be change by align_transform)
+        depth_resolution = Realsense.RESOLUTION_720 if fps <= 30 else Realsense.RESOLUTION_480
         if self._depth_enabled:
-            # Use max resolution that can handle the fps for depth (will be change by align_transform)
-            depth_resolution = Realsense.RESOLUTION_720 if fps <= 30 else Realsense.RESOLUTION_480
             config.enable_stream(rs.stream.depth, depth_resolution[0], depth_resolution[1], rs.format.z16, fps)
+
+        if self._confidence_enabled:
+            config.enable_stream(rs.stream.infrared, 1, depth_resolution[0], depth_resolution[1], rs.format.y8, fps)
+            config.enable_stream(rs.stream.infrared, 2, depth_resolution[0], depth_resolution[1], rs.format.y8, fps)
 
         # Avoid having to reconnect the USB cable, see https://github.com/IntelRealSense/librealsense/issues/6628#issuecomment-646558144
         ctx = rs.context()
@@ -137,6 +148,12 @@ class Realsense(RGBDCamera):
         if self.hole_filling_enabled:
             self._depth_frame = self.hole_filling.process(self._depth_frame)
 
+        if not self._confidence_enabled:
+            return
+
+        self._infrared_frame_1 = aligned_frames.get_infrared_frame(1)
+        self._infrared_frame_2 = aligned_frames.get_infrared_frame(2)
+
     def _retrieve_rgb_image(self) -> NumpyFloatImageType:
         image = self._retrieve_rgb_image_as_int()
         return ImageConverter.from_numpy_int_format(image).image_in_numpy_format
@@ -149,20 +166,68 @@ class Realsense(RGBDCamera):
         return image
 
     def _retrieve_depth_map(self) -> NumpyDepthMapType:
+        if not self._depth_enabled:
+            raise RuntimeError("Cannot retrieve depth data if depth is disabled")
         frame = self._depth_frame
         image = np.asanyarray(frame.get_data()).astype(np.float32)
         return image * self.depth_factor
 
     def _retrieve_depth_image(self) -> NumpyIntImageType:
+        if not self._depth_enabled:
+            raise RuntimeError("Cannot retrieve depth data if depth is disabled")
         frame = self._depth_frame
         frame_colorized = self.colorizer.colorize(frame)
         image = np.asanyarray(frame_colorized.get_data())  # this is uint8 with 3 channels
         return image
 
+    def _retrieve_confidence_map(self) -> NumpyConfidenceMapType:
+        # Compute confidence map based on the disparity between the two IR images.
+        if not self._confidence_enabled:
+            raise RuntimeError("Cannot retrieve confidence data if confidence is disabled")
+        if not isinstance(self._composite_frame, rs.composite_frame):
+            raise RuntimeError("_grab_images must be called before retrieving images")
+        ir1_frame = self._infrared_frame_1
+        ir2_frame = self._infrared_frame_2
+
+        # Convert images to numpy
+        ir1 = np.asanyarray(ir1_frame.get_data())
+        ir2 = np.asanyarray(ir2_frame.get_data())
+
+        # default values for SGBM according to OpenCV docs
+        wls_lambda = 8000.0
+        wls_sigma = 1.5
+        if not hasattr(self, "_stereo_sgbm"):
+            max_disp = 160  # must be divisible by 16
+            window_size = 3
+            p1 = 216  # 24 * window_size ** 2
+            p2 = 864  # 96 * window_size ** 2
+            pre_filter_cap = 63
+
+            self._stereo_sgbm = cv2.StereoSGBM.create(
+                minDisparity=0,
+                numDisparities=max_disp,
+                blockSize=window_size,
+                P1=p1,
+                P2=p2,
+                preFilterCap=pre_filter_cap,
+                mode=cv2.StereoSGBM_MODE_SGBM_3WAY,
+            )
+
+        left_matcher = self._stereo_sgbm
+        wls_filter = cv2.ximgproc.createDisparityWLSFilter(left_matcher)
+        right_matcher = cv2.ximgproc.createRightMatcher(left_matcher)
+        left_disp = left_matcher.compute(ir1, ir2).astype(np.float32) / 16.0
+        right_disp = right_matcher.compute(ir2, ir1).astype(np.float32) / 16.0
+        wls_filter.setLambda(wls_lambda)
+        wls_filter.setSigmaColor(wls_sigma)
+        wls_filter.filter(left_disp, ir1, disparity_map_right=right_disp)
+        confidence_map = wls_filter.getConfidenceMap()
+
+        return confidence_map / 255.0
+
 
 if __name__ == "__main__":
     import airo_camera_toolkit.cameras.manual_test_hw as test
-    import cv2
 
     camera = Realsense(fps=30, resolution=Realsense.RESOLUTION_1080, enable_hole_filling=True)